move tests for ModelCriteria::populateRelation() into own file
