+++ tests/Propel/Tests/Runtime/ActiveQuery/ModelCriteriaTest.php
@@ -19,6 +19,7 @@ use Propel\Runtime\Exception\ClassNotFoundException;
 use Propel\Runtime\Exception\EntityNotFoundException;
 use Propel\Runtime\Exception\PropelException;
 use Propel\Runtime\Exception\UnexpectedValueException;
+use Propel\Runtime\Map\ColumnMap;
 use Propel\Runtime\Map\Exception\RelationNotFoundException;
 use Propel\Runtime\Propel;
 use Propel\Runtime\Util\PropelModelPager;
@@ -1293,342 +1294,6 @@ class ModelCriteriaTest extends BookstoreTestBase
 
         $joins = $c->getJoins();
         $this->assertEquals($joins['Author'], $c->getJoin('Author'), 'getJoin() returns a specific Join from the ModelCriteria');
-    }
-
-    /**
-     * @return void
-     */
-    public function testWith()
-    {
-        $c = new ModelCriteria('bookstore', 'Propel\Tests\Bookstore\Book');
-        $c->join('Propel\Tests\Bookstore\Book.Author');
-        $c->populateJoinedRelation('Author');
-        $withs = $c->getRelatedModelsToPopulate();
-        $this->assertTrue(array_key_exists('Author', $withs), 'with() adds an entry to the internal list of Withs');
-        $this->assertInstanceOf('Propel\Runtime\ActiveQuery\ModelWith', $withs['Author'], 'with() references the ModelWith object');
-    }
-
-    /**
-     * @return void
-     */
-    public function testWithThrowsExceptionWhenJoinLacks()
-    {
-        $this->expectException(UnknownRelationException::class);
-
-        $c = new ModelCriteria('bookstore', 'Propel\Tests\Bookstore\Book');
-        $c->populateJoinedRelation('Propel\Tests\Bookstore\Author');
-    }
-
-    /**
-     * @return void
-     */
-    public function testWithAlias()
-    {
-        $c = new ModelCriteria('bookstore', 'Propel\Tests\Bookstore\Book');
-        $c->join('Propel\Tests\Bookstore\Book.Author a');
-        $c->populateJoinedRelation('a');
-        $withs = $c->getRelatedModelsToPopulate();
-        $this->assertTrue(array_key_exists('a', $withs), 'with() uses the alias for the index of the internal list of Withs');
-    }
-
-    /**
-     * @return void
-     */
-    public function testWithThrowsExceptionWhenNotUsingAlias()
-    {
-        $this->expectException(UnknownRelationException::class);
-
-        $c = new ModelCriteria('bookstore', 'Propel\Tests\Bookstore\Book');
-        $c->join('Propel\Tests\Bookstore\Book.Author a');
-        $c->populateJoinedRelation('Propel\Tests\Bookstore\Author');
-    }
-
-    /**
-     * @return void
-     */
-    public function testWithAddsSelectColumns()
-    {
-        $c = new ModelCriteria('bookstore', 'Propel\Tests\Bookstore\Book');
-        $c->join('Propel\Tests\Bookstore\Book.Author');
-        $c->populateJoinedRelation('Author');
-        $expectedColumns = [
-            BookTableMap::COL_ID,
-            BookTableMap::COL_TITLE,
-            BookTableMap::COL_ISBN,
-            BookTableMap::COL_PRICE,
-            BookTableMap::COL_PUBLISHER_ID,
-            BookTableMap::COL_AUTHOR_ID,
-            AuthorTableMap::COL_ID,
-            AuthorTableMap::COL_FIRST_NAME,
-            AuthorTableMap::COL_LAST_NAME,
-            AuthorTableMap::COL_EMAIL,
-            AuthorTableMap::COL_AGE,
-        ];
-        $this->assertEquals($expectedColumns, $c->getSelectColumns(), 'with() adds the columns of the related table');
-    }
-
-    /**
-     * @return void
-     */
-    public function testWithAliasAddsSelectColumns()
-    {
-        $c = new ModelCriteria('bookstore', 'Propel\Tests\Bookstore\Book');
-        $c->join('Propel\Tests\Bookstore\Book.Author a');
-        $c->populateJoinedRelation('a');
-        $expectedColumns = [
-            BookTableMap::COL_ID,
-            BookTableMap::COL_TITLE,
-            BookTableMap::COL_ISBN,
-            BookTableMap::COL_PRICE,
-            BookTableMap::COL_PUBLISHER_ID,
-            BookTableMap::COL_AUTHOR_ID,
-            'a.id',
-            'a.first_name',
-            'a.last_name',
-            'a.email',
-            'a.age',
-        ];
-        $this->assertEquals($expectedColumns, $c->getSelectColumns(), 'with() adds the columns of the related table');
-    }
-
-    /**
-     * @return void
-     */
-    public function testWithAddsSelectColumnsOfMainTable()
-    {
-        $c = new ModelCriteria('bookstore', 'Propel\Tests\Bookstore\Book');
-        $c->join('Propel\Tests\Bookstore\Book.Author');
-        $c->populateJoinedRelation('Author');
-        $expectedColumns = [
-            BookTableMap::COL_ID,
-            BookTableMap::COL_TITLE,
-            BookTableMap::COL_ISBN,
-            BookTableMap::COL_PRICE,
-            BookTableMap::COL_PUBLISHER_ID,
-            BookTableMap::COL_AUTHOR_ID,
-            AuthorTableMap::COL_ID,
-            AuthorTableMap::COL_FIRST_NAME,
-            AuthorTableMap::COL_LAST_NAME,
-            AuthorTableMap::COL_EMAIL,
-            AuthorTableMap::COL_AGE,
-        ];
-        $this->assertEquals($expectedColumns, $c->getSelectColumns(), 'with() adds the columns of the main table if required');
-    }
-
-    /**
-     * @return void
-     */
-    public function testWithAliasAddsSelectColumnsOfMainTable()
-    {
-        $c = new ModelCriteria('bookstore', 'Propel\Tests\Bookstore\Book');
-        $c->setModelAlias('b', true);
-        $c->join('b.Author a');
-        $c->populateJoinedRelation('a');
-        $expectedColumns = [
-            'b.id',
-            'b.title',
-            'b.isbn',
-            'b.price',
-            'b.publisher_id',
-            'b.author_id',
-            'a.id',
-            'a.first_name',
-            'a.last_name',
-            'a.email',
-            'a.age',
-        ];
-        $this->assertEquals($expectedColumns, $c->getSelectColumns(), 'with() adds the columns of the main table with an alias if required');
-    }
-
-    /**
-     * @return void
-     */
-    public function testWithOneToManyAddsSelectColumns()
-    {
-        $c = new ModelCriteria('bookstore', 'Propel\Tests\Bookstore\Author');
-        $c->leftJoin('Propel\Tests\Bookstore\Author.Book');
-        $c->populateJoinedRelation('Book');
-        $expectedColumns = [
-            AuthorTableMap::COL_ID,
-            AuthorTableMap::COL_FIRST_NAME,
-            AuthorTableMap::COL_LAST_NAME,
-            AuthorTableMap::COL_EMAIL,
-            AuthorTableMap::COL_AGE,
-            BookTableMap::COL_ID,
-            BookTableMap::COL_TITLE,
-            BookTableMap::COL_ISBN,
-            BookTableMap::COL_PRICE,
-            BookTableMap::COL_PUBLISHER_ID,
-            BookTableMap::COL_AUTHOR_ID,
-        ];
-        $this->assertEquals($expectedColumns, $c->getSelectColumns(), 'with() adds the columns of the related table even in a one-to-many relationship');
-    }
-
-    public static function PopulateDataProvider(): array
-    {
-        return [
-            ['joinWith', Criteria::INNER_JOIN], 
-            ['populateRelation', Criteria::LEFT_JOIN], 
-        ];
-    }
-
-    /**
-     * @return void
-     */
-    #[DataProvider('PopulateDataProvider')]
-    public function testJoinWith(string $populateFunctionName, string $expectedJoinType)
-    {
-        $c = new ModelCriteria('bookstore', 'Propel\Tests\Bookstore\Book');
-        $c->$populateFunctionName('Propel\Tests\Bookstore\Book.Author');
-        $expectedColumns = [
-            BookTableMap::COL_ID,
-            BookTableMap::COL_TITLE,
-            BookTableMap::COL_ISBN,
-            BookTableMap::COL_PRICE,
-            BookTableMap::COL_PUBLISHER_ID,
-            BookTableMap::COL_AUTHOR_ID,
-            AuthorTableMap::COL_ID,
-            AuthorTableMap::COL_FIRST_NAME,
-            AuthorTableMap::COL_LAST_NAME,
-            AuthorTableMap::COL_EMAIL,
-            AuthorTableMap::COL_AGE,
-        ];
-        $this->assertEquals($expectedColumns, $c->getSelectColumns(), "$populateFunctionName() adds the join");
-        $joins = $c->getJoins();
-        $join = $joins['Author'];
-        $this->assertEquals($expectedJoinType, $join->getJoinType(), "$populateFunctionName() adds an $expectedJoinType by default");
-    }
-
-    /**
-     * @return void
-     */
-    public function testJoinWithType()
-    {
-        $c = new ModelCriteria('bookstore', 'Propel\Tests\Bookstore\Book');
-        $c->populateRelation('Propel\Tests\Bookstore\Book.Author', Criteria::LEFT_JOIN);
-        $joins = $c->getJoins();
-        $join = $joins['Author'];
-        $this->assertEquals(Criteria::LEFT_JOIN, $join->getJoinType(), 'joinWith() accepts a join type as second parameter');
-    }
-
-    /**
-     * @return void
-     */
-    public function testJoinWithAlias()
-    {
-        $c = new ModelCriteria('bookstore', 'Propel\Tests\Bookstore\Book');
-        $c->populateRelation('Propel\Tests\Bookstore\Book.Author a');
-        $expectedColumns = [
-            BookTableMap::COL_ID,
-            BookTableMap::COL_TITLE,
-            BookTableMap::COL_ISBN,
-            BookTableMap::COL_PRICE,
-            BookTableMap::COL_PUBLISHER_ID,
-            BookTableMap::COL_AUTHOR_ID,
-            'a.id',
-            'a.first_name',
-            'a.last_name',
-            'a.email',
-            'a.age',
-        ];
-        $this->assertEquals($expectedColumns, $c->getSelectColumns(), 'joinWith() adds the join with the alias');
-    }
-
-    /**
-     * @return void
-     */
-    public function testJoinWithSeveral()
-    {
-        $c = new ModelCriteria('bookstore', 'Propel\Tests\Bookstore\Review');
-        $c->populateRelation('Review.Book');
-        $c->populateRelation('Propel\Tests\Bookstore\Book.Author');
-        $c->populateRelation('Book.Publisher');
-        $expectedColumns = [
-            ReviewTableMap::COL_ID,
-            ReviewTableMap::COL_REVIEWED_BY,
-            ReviewTableMap::COL_REVIEW_DATE,
-            ReviewTableMap::COL_RECOMMENDED,
-            ReviewTableMap::COL_STATUS,
-            ReviewTableMap::COL_BOOK_ID,
-            BookTableMap::COL_ID,
-            BookTableMap::COL_TITLE,
-            BookTableMap::COL_ISBN,
-            BookTableMap::COL_PRICE,
-            BookTableMap::COL_PUBLISHER_ID,
-            BookTableMap::COL_AUTHOR_ID,
-            AuthorTableMap::COL_ID,
-            AuthorTableMap::COL_FIRST_NAME,
-            AuthorTableMap::COL_LAST_NAME,
-            AuthorTableMap::COL_EMAIL,
-            AuthorTableMap::COL_AGE,
-            PublisherTableMap::COL_ID,
-            PublisherTableMap::COL_NAME,
-        ];
-        $this->assertEquals($expectedColumns, $c->getSelectColumns(), 'joinWith() adds the with');
-        $joins = $c->getJoins();
-        $expectedJoinKeys = ['Book', 'Author', 'Publisher'];
-        $this->assertEquals($expectedJoinKeys, array_keys($joins), 'joinWith() adds the join');
-    }
-
-    /**
-     * @return void
-     */
-    public function testJoinWithTwice()
-    {
-        $c = new ModelCriteria('bookstore', 'Propel\Tests\Bookstore\Book');
-        $c->join('Propel\Tests\Bookstore\Book.Review');
-        $c->populateRelation('Propel\Tests\Bookstore\Book.Author');
-        $c->populateRelation('Propel\Tests\Bookstore\Book.Review');
-        $expectedColumns = [
-            BookTableMap::COL_ID,
-            BookTableMap::COL_TITLE,
-            BookTableMap::COL_ISBN,
-            BookTableMap::COL_PRICE,
-            BookTableMap::COL_PUBLISHER_ID,
-            BookTableMap::COL_AUTHOR_ID,
-            AuthorTableMap::COL_ID,
-            AuthorTableMap::COL_FIRST_NAME,
-            AuthorTableMap::COL_LAST_NAME,
-            AuthorTableMap::COL_EMAIL,
-            AuthorTableMap::COL_AGE,
-            ReviewTableMap::COL_ID,
-            ReviewTableMap::COL_REVIEWED_BY,
-            ReviewTableMap::COL_REVIEW_DATE,
-            ReviewTableMap::COL_RECOMMENDED,
-            ReviewTableMap::COL_STATUS,
-            ReviewTableMap::COL_BOOK_ID,
-        ];
-        $this->assertEquals($expectedColumns, $c->getSelectColumns(), 'joinWith() adds the with');
-        $joins = $c->getJoins();
-        $expectedJoinKeys = ['Review', 'Author'];
-        $this->assertEquals($expectedJoinKeys, array_keys($joins), 'joinWith() adds the join');
-    }
-
-    /**
-     * @return void
-     */
-    public function testPopulateWithVirtualColumns()
-    {
-        $book = (new Book)->fromArray(['Title' => 'City of Glass', 'ISBN' => '1234']);
-        $author = (new Author())->fromArray(['FirstName' => 'Paul', 'LastName' => 'Auster'])->addBook($book);
-        $author->save();
-        AuthorTableMap::clearInstancePool();
-        BookTableMap::clearInstancePool();
-
-        $reloadedAuthor = AuthorQuery::create()
-            ->populateBook()
-            ->addAsColumn('leAuthorId', 'author.id')
-            ->filterById($author->getId())
-            ->find()->get(0);
-
-        $expectedAuthorArray = array_merge($author->toArray(), ['leAuthorId' => $author->getId()]);
-        $this->assertSame($expectedAuthorArray, $reloadedAuthor->toArray());
-        $this->assertSame(['leAuthorId' => $author->getId()], $reloadedAuthor->getVirtualColumns());
-
-        $books = $this->getObjectPropertyValue($reloadedAuthor, 'collBooks');
-        $this->assertCount(1, $books);
-        $this->assertSame($book->toArray(), $books[0]->toArray());
-
     }
 
     public static function conditionsForTestWithColumn()
